% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot Accidents for Provided Year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A state number; accepts text or number but is coerced
to number}

\item{years}{A year; accepts text or number but is coerced to number}
}
\value{
A plot of accidents in the state from the provided year
}
\description{
Reads a files specified by a year value and plots accidents in the state
Requires the following packages: \code{readr}, \code{dplyr}, \code{tidyr},
\code{maps}, \code{graphics}, \code{magrittr}
If a state number is not contained in the file, this function will throw
an error.
If no accidents are available for the state provided, this function will
print a message and return \code{NULL}
}
\examples{
#These all produce the same output
fars_map_state(2013)
fars_map_state('2013')

}
