% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads an Accident File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename}
}
\value{
A \code{tibble} containing the read-in data
}
\description{
This is a simple function that reads in a file using \code{readr::read_csv} and
converts it to a \code{tibble}.
Requires the following packages: \code{readr}, \code{dplyr}
}
\examples{
fars_read("accident_2013.csv")

}
