% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads Datasets for Provided Year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years, presented as text or numbers}
}
\value{
A list of data sets.
}
\description{
Reads a list of files specified by a set of year values, returning a list
of data sets
Requires the following packages: \code{readr}, \code{dplyr},\code{magrittr}
}
\examples{
#These all produce the same output
fars_read_years(c(2013,2014))
fars_read_years(c('2013','2014'))
fars_read_years(list(2013,2014))
fars_read_years(list('2013','2014'))

}
