% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count Accidents by Provided Year(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years, presented as text or numbers}
}
\value{
A table with one row per month and columns corresponding to the
month and each provided year, where the latter columns contain the number
of accidents in that month in that year.
}
\description{
Reads a list of files specified by a set of year values, counts the
number of records in each year-month combination, and pivots to
wide format with separate columns per year
Requires the following packages: \code{readr}, \code{dplyr}, \code{tidyr},
\code{magrittr}
}
\examples{
\dontrun{
#These all produce the same output
fars_summarize_years(c(2013,2014))
fars_summarize_years(c('2013','2014'))
fars_summarize_years(list(2013,2014))
fars_summarize_years(list('2013','2014'))
}

}
