% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes a filename for use in \code{fars_read}}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year value or vector of such. Will be coerced to an integer.}
}
\value{
A text string or vector of such corresponding to file names.
}
\description{
This is a simple function that takes in a year (e.g., 2013) or vector of
years and creates a text string to use as a filename input to
\code{\link{fars_read}}.
}
\examples{
make_filename(2013)
make_filename('2013') #NOTE: coerced to integer

}
